requirements.txt
selenium
pandas
aiogram==3.18.0
python-dotenv


